requirements.txt
asgiref==3.11.0
black==25.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cloudinary==1.44.1
Django==6.0
django-cloudinary-storage==0.3.0
djangorestframework==3.16.1
drf-yasg==1.21.11
idna==3.11
inflection==0.5.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
ruff==0.14.10
six==1.17.0
sqlparse==0.5.5
uritemplate==4.2.0
urllib3==2.6.2
